reaction 함수도 역할 분리 똑같이 적용
